[tune] Move experiment state/checkpoint/resume management into a separate file (#32457)
Experiment state management is currently convoluted.
We keep track of many duplicate variables, e.g. local/remote checkpoint dirs and syncers.
The resume/syncing logic also takes up a lot of space in the trial runner.

Saving and restoring experiment state is orthogonal to the actual trial lifecycle logic, thus it makes sense to separate this out. In the same go, I've removed a lot of duplicated state and simplified some APIs that will also make it easier to test the experiment state component separately.
